use node APIs instead of nextjs
